# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "Laravel API OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Laravel API Server"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "USER"
                ],
                "summary": "Register a new user",
                "operationId": "596169d03e90fa336fb3e1aa9bfe7367",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password, password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "a@b.c"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "johns-iphone"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "token"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "USER"
                ],
                "summary": "Login a user",
                "operationId": "d1572a3de62d3f37c8c83d0f0bc13c88",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password, password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "a@b.c"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "johns-iphone"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User logged in successfully"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "token"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "USER"
                ],
                "summary": "Get curent user",
                "operationId": "2c4d73bb274ed7149a33efcac3710d81",
                "responses": {
                    "200": {
                        "description": "User is logged",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "302": {
                        "description": "Not Authed"
                    }
                }
            }
        },
        "/logout": {
            "patch": {
                "tags": [
                    "USER"
                ],
                "summary": "logout a user",
                "operationId": "f784421249244c68ddce8343f58a5f2c",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "bearerAuth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User logged out successfully"
                    },
                    "403": {
                        "description": "Auth error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "USER"
                ],
                "summary": "get all users paginated",
                "operationId": "657f9241eff9e55dae11f1160045c612",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "bearerAuth"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "allowEmptyValue": true,
                        "example": "michel"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginatedResult"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/User"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Auth error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{userId}": {
            "delete": {
                "tags": [
                    "USER"
                ],
                "summary": "Delete user",
                "operationId": "cd2d3aa3b5a7385411927c7ad45a1e0e",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "bearerAuth"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Id Of selected user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sensor deleted succesfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Not Allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "PaginatedResult": {
                "title": "Pagination exemple",
                "properties": {
                    "total": {
                        "type": "integer"
                    },
                    "current_page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "next_page_url": {
                        "type": "string"
                    },
                    "prev_page_url": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "schema of user",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}